requirements.txt
flask
flask_mysqldb
PyJWT